Fixing comparison of narrow type with wide type in loop condition

+++ ChiTech/ChiMesh/SweepUtilities/FLUDS/FLUDS_alphapass_01_local_inc_mapping.cc
@@ -25,7 +25,7 @@ LocalIncidentMapping(chi_mesh::Cell *cell,
   //=================================================== Loop over faces
   //           INCIDENT                                 but process
   //                                                    only incident faces
-  for (short f=0; f < cell->faces.size(); f++)
+  for (size_t f=0; f < cell->faces.size(); f++)
   {
     CellFace& face = cell->faces[f];
     double     mu  = face.normal.Dot(spds->omega);
@@ -50,7 +50,7 @@ LocalIncidentMapping(chi_mesh::Cell *cell,
         int  ass_f_counter=-1;
 
         int out_f = -1;
-        for (short af=0; af < adj_cell->faces.size(); af++)
+        for (size_t af=0; af < adj_cell->faces.size(); af++)
         {
           double mur = adj_cell->faces[af].normal.Dot(spds->omega);
